requirements.txt
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==3.0.3
apache-airflow-core==3.0.3
apache-airflow-providers-common-compat==1.7.2
apache-airflow-providers-common-io==1.6.1
apache-airflow-providers-common-sql==1.27.3
apache-airflow-providers-databricks==7.6.0
apache-airflow-providers-fab==2.3.0
apache-airflow-providers-smtp==2.1.1
apache-airflow-providers-standard==1.4.1
apache-airflow-task-sdk==1.0.3
apispec==6.8.2
argcomplete==3.6.2
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
backports-datetime-fromisoformat==2.0.3
backports.strenum==1.3.1
blinker==1.9.0
cachelib==0.13.0
cadwyn==5.4.3
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
connexion==2.14.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==42.0.8
databricks-sql-connector==4.0.5
databricks-sqlalchemy==1.0.5
Deprecated==1.2.18
dill==0.3.1.1
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
Flask==2.2.5
Flask-AppBuilder==4.6.3
Flask-Babel==2.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.7.0
fsspec==2025.5.1
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.10
importlib_metadata==8.4.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lazy-object-proxy==1.11.0
libcst==1.8.2
limits==5.4.0
linkify-it-py==2.0.3
lockfile==0.12.2
lz4==4.4.4
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
mergedeep==1.3.4
methodtools==0.4.7
more-itertools==10.7.0
msgspec==0.19.0
multidict==6.6.3
numpy==2.2.6
oauthlib==3.3.1
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
ordered-set==4.1.0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pendulum==3.1.0
pluggy==1.6.0
prison==0.2.1
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
pyarrow==16.1.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
retryhttp==1.3.3
rich==13.9.4
rich-argparse==1.7.1
rich-toolkit==0.14.8
rignore==0.5.1
rpds-py==0.26.0
sentry-sdk==2.32.0
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
starlette==0.46.2
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
thrift==0.20.0
tomli==2.2.1
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==14.2
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.2.1
yarl==1.20.1
zipp==3.23.0
